moved an exception to another package
