+++ src/main/java/com/project/app/open_school_t1/exception/TaskNotFoundException.java
@@ -1,4 +1,4 @@
-package com.project.app.open_school_t1.util;
+package com.project.app.open_school_t1.exception;
 
 public class TaskNotFoundException extends RuntimeException {
     public TaskNotFoundException(String message) {
